Make controller naming consistent
