+++ app/Http/Requests/StoreAdvertRequest.php
@@ -4,7 +4,7 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateJobApplicationRequest extends FormRequest
+class StoreAdvertRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
